+++ tests/Event/ReadModel/JSONLD/EventLDProjectorTest.php
@@ -19,6 +19,10 @@ use CultuurNet\UDB3\Event\Events\ChildrenOnlyUpdated;
 use CultuurNet\UDB3\Event\Events\DeparturePlacesUpdated;
 use CultuurNet\UDB3\Event\Events\BirthdateRangeDeleted;
 use CultuurNet\UDB3\Event\Events\BirthdateRangeUpdated;
+use CultuurNet\UDB3\Media\Image;
+use CultuurNet\UDB3\Media\Properties\Description as MediaDescription;
+use CultuurNet\UDB3\Media\Properties\MIMEType;
+use CultuurNet\UDB3\Model\ValueObject\MediaObject\CopyrightHolder;
 use CultuurNet\UDB3\Model\ValueObject\Web\Url;
 use CultuurNet\UDB3\Model\ValueObject\Web\Urls;
 use CultuurNet\UDB3\Event\Events\FaqsUpdated;
@@ -586,6 +590,102 @@ class EventLDProjectorTest extends OfferLDProjectorTestBase
         $expectedJsonLD->completeness = 60;
 
         $this->assertEquals($expectedJsonLD, $body);
+    }
+
+    /**
+     * @test
+     */
+    public function it_projects_the_adding_of_an_image(): void
+    {
+        $id = 'foo';
+        $imageId = new Uuid('de305d54-75b4-431b-adb2-eb6b9e546014');
+        $description = new MediaDescription('Some description.');
+        $copyrightHolder = new CopyrightHolder('Dirk Dirkington');
+        $type = new MIMEType('image/png');
+        $location = new Url('http://foo.bar/media/de305d54-75b4-431b-adb2-eb6b9e546014.png');
+        $language = new Language('en');
+
+        $image = new Image($imageId, $type, $description, $copyrightHolder, $location, $language);
+        $eventClass = $this->getEventClass('ImageAdded');
+        $imageAdded = new $eventClass($id, $image);
+
+        $initialDocument = new JsonDocument($id);
+        $this->documentRepository->save($initialDocument);
+
+        $expectedBody = (object)[
+            'image' => 'http://foo.bar/media/de305d54-75b4-431b-adb2-eb6b9e546014.png',
+            'mediaObject' => [
+                (object)[
+                    '@id' => 'http://example.com/entity/de305d54-75b4-431b-adb2-eb6b9e546014',
+                    '@type' => 'schema:ImageObject',
+                    'id' => 'de305d54-75b4-431b-adb2-eb6b9e546014',
+                    'contentUrl' => 'http://foo.bar/media/de305d54-75b4-431b-adb2-eb6b9e546014.png',
+                    'thumbnailUrl' => 'http://foo.bar/media/de305d54-75b4-431b-adb2-eb6b9e546014.png',
+                    'description' => $description->toString(),
+                    'copyrightHolder' => $copyrightHolder->toString(),
+                    'inLanguage' => 'en',
+                ],
+            ],
+            'modified' => $this->recordedOn->toString(),
+            'playhead' => 1,
+            'completeness' => 7,
+        ];
+
+        $body = $this->project($imageAdded, $id, null, $this->recordedOn->toBroadwayDateTime());
+
+        $this->assertEquals($expectedBody, $body);
+    }
+
+    /**
+     * @test
+     */
+    public function it_projects_the_editing_of_an_image(): void
+    {
+        $id = 'foo';
+        $imageId = new Uuid('de305d54-75b4-431b-adb2-eb6b9e546014');
+        $description = 'Some description.';
+        $copyrightHolder = new CopyrightHolder('Dirk Dirkington');
+        $eventClass = $this->getEventClass('ImageUpdated');
+        $imageUpdated = new $eventClass($id, $imageId->toString(), $description, $copyrightHolder->toString());
+
+        $initialDocument = new JsonDocument(
+            $id,
+            Json::encode([
+                'mediaObject' => [
+                    [
+                        '@id' => 'http://example.com/entity/de305d54-75b4-431b-adb2-eb6b9e546014',
+                        '@type' => 'schema:ImageObject',
+                        'contentUrl' => 'http://foo.bar/media/de305d54-75b4-431b-adb2-eb6b9e546014.png',
+                        'thumbnailUrl' => 'http://foo.bar/media/de305d54-75b4-431b-adb2-eb6b9e546014.png',
+                        'description' => 'olddescription',
+                        'copyrightHolder' => 'oldcopyrightHolder',
+                        'inLanguage' => 'en',
+                    ],
+                ],
+            ])
+        );
+        $this->documentRepository->save($initialDocument);
+
+        $expectedBody = (object)[
+            'mediaObject' => [
+                (object)[
+                    '@id' => 'http://example.com/entity/de305d54-75b4-431b-adb2-eb6b9e546014',
+                    '@type' => 'schema:ImageObject',
+                    'contentUrl' => 'http://foo.bar/media/de305d54-75b4-431b-adb2-eb6b9e546014.png',
+                    'thumbnailUrl' => 'http://foo.bar/media/de305d54-75b4-431b-adb2-eb6b9e546014.png',
+                    'description' => (string) $description,
+                    'copyrightHolder' => $copyrightHolder->toString(),
+                    'inLanguage' => 'en',
+                ],
+            ],
+            'modified' => $this->recordedOn->toString(),
+            'playhead' => 1,
+            'completeness' => 7,
+        ];
+
+        $body = $this->project($imageUpdated, $id, null, $this->recordedOn->toBroadwayDateTime());
+
+        $this->assertEquals($expectedBody, $body);
     }
 
     /**
